# adds PREG_NO_UTF8_CHECK flag