… ext/pcre/tests/pcre.constants.phpt