Hi,
This PR is related to [Bug #72685](https://bugs.php.net/bug.php?id=72685).
It adds a `PREG_NO_UTF8_CHECK` flag to `preg_match`, `preg_match_all` and `preg_split` functions to disable the implicit UTF-8 validity checking performed by PCRE by default when the `/u` option is enabled.

Without the ability to disable this check, matching repeatedly against the same input in utf-8 mode yields quadratic performance , instead of linear.

I'm targetting this branch because it fixes a bug and it is a BC change, but it works on master too.
This is my first PR to php-src, so feel free to point me in the right direction if needed.

Thanks.

Following is a small PHP script that can be ran against this branch to test the new behavior:

```php
function do_test ($input_size, $flags = 0) {
  $str = str_repeat('a', $input_size);

  $start = microtime(true);
  $pos = 0;
  while(preg_match('/\G\w/', $str, $m, $flags, $pos)) ++$pos;
  $end = microtime(true);
  echo '>>> NO u flag: ', number_format(($end - $start)*1000, 6), 'ms', PHP_EOL;

  $start = microtime(true);
  $pos = 0;
  while(preg_match('/\G\w/u', $str, $m, $flags, $pos)) ++$pos;
  $end = microtime(true);
  echo '>>> WITH u flag: ', number_format(($end - $start)*1000, 6), 'ms', PHP_EOL;
}

echo 'WITHOUT PREG_NO_UTF8_CHECK', PHP_EOL;
echo 'Input size = 1e4 ', PHP_EOL;
do_test(1e4, 0);
echo 'Input size = 1e5 ', PHP_EOL;
do_test(1e5, 0);
echo str_repeat('-', 80), PHP_EOL;
echo 'WITH PREG_NO_UTF8_CHECK', PHP_EOL;
echo 'Input size = 1e4 ', PHP_EOL;
do_test(1e4, PREG_NO_UTF8_CHECK);
echo 'Input size = 1e5 ', PHP_EOL;
do_test(1e5, PREG_NO_UTF8_CHECK);
```

Sample output:
```
WITHOUT PREG_NO_UTF8_CHECK
Input size = 1e4
>>> NO u flag:    14.562130ms
>>> WITH u flag: 243.849993ms
Input size = 1e5
>>> NO u flag:      111.440897ms
>>> WITH u flag: 23,129.899025ms
--------------------------------------------------------------------------------
WITH PREG_NO_UTF8_CHECK
Input size = 1e4
>>> NO u flag:   11.296988ms
>>> WITH u flag: 11.363029ms
Input size = 1e5
>>> NO u flag:   110.867023ms
>>> WITH u flag: 111.830950ms
```