@weltling 2016-07-27
@ju1ius thanks for the report. I would suggest to turn your PHP code into the test and add to the PR. I also think it is required to add tests with invalid UTF-8 strings.

While this option can be of course useful, it looks dangerous to be allowed just as is. The PCRE man page warns explicitly about possible bad consequences for the case this option is used incorrectly. Not sure, maybe hashing the subject string before were an option, if utf-8 check is disabled. Or some other way to ensure same subject string is passed over again.

Thanks.