@ju1ius 2016-07-27
>  I also think it is required to add tests with invalid UTF-8 strings.

Yep, just doing that ATM :wink: