@ju1ius 2016-07-27
> While this option can be of course useful, it looks dangerous to be allowed just as is.
> Not sure, maybe hashing the subject string before were an option, if utf-8 check is disabled.

Wouldn't hashing the string on each invocation kind of defeat the point ?

Also, no pun intended but PHP already has an history of trying to be automagically safe, failing at that, thus creating security holes  and (what's worth) long lasting bad habits among it's community (remember magic_quotes, safe_mode, et al.)

We already have a function to check for UTF-8 validity: [`mb_check_encoding`](http://fr2.php.net/manual/en/function.mb-check-encoding.php)! :tada:

so IMO the right solution would be to document this flag with a big, bold, yellow warning like:

> Use this flag when you need to match repeatedly against the same input in utf-8 mode.
> Be sure however to call `mb_check_encoding($my_input, 'utf-8')` before matching, otherwise things will blow-up and a fairy will die.

And include a small code sample like:

```php
if (!mb_check_encoding($input, 'utf-8')) {
  throw new \InvalidArgumentException('Invalid UTF-8 string');
}
// Now parse all the things a million times with
$pos = 0;
while(preg_match('/\G\s*(\w+)\s*/u', $input, $matches, PREG_NO_UTF8_CHECK, $pos)) {
  $pos += strlen($matches[0]);
  $my_word = $matches[1];
}
```

What do you think? :wink: