@ju1ius 2016-07-27
@weltling I just added a few tests based on what I could find inside the `ext/pcre/tests` directory

> I would suggest to turn your PHP code into the test and add to the PR.

I'm not sure how to do that, the above code is more a benchmark. Can you point me to existing performance tests ?