@weltling 2016-07-27
Yeah, it is a clear that the validation has to happen before just once. What I'm talking about is the case where no validation happened before. PCRE promises undefined behavior or even a crash. Imho not a good idea just opening a rabbit hole without any defense. Documentation is a good thing, but maybe more could be done. Hashing was just the first thing that came into my mind - it iis supposed to be faster than utf-8 check. A smarter idea is welcome :)

Thanks.