@weltling 2016-07-27
With the benchmark test - you can pick a criteria. Say the time difference is expected to be not more that X ms, etc. Or even seconds. Just to be sure that even the rough expectations are ensured.

Thanks.