@staabm 2016-07-27
could php-src somehow remember which strings already have been "validated" and set this flag automatically instead of defering this decision to userland?