@nikic 2016-07-27
Just like @weltling I'm apprehensive about adding this. The documentation is pretty explicit on the point:

 > When PCRE_NO_UTF8_CHECK is set, the effect of passing an invalid string as a subject or an invalid value of startoffset is undefined. Your program may crash or loop.

We *cannot* expose options that cause memory unsafety. If executing PHP code can cause memory errors (-so), this is always a bug in PHP and not a problem with the user doing something they're not supposed to do.

@staabm In principle possible. PCRE can keep an HT from str addr to str of validated strings. This would allow skipping validation if the same string instance is passed. It's a tradeoff though, because this would be keeping alive strings. Would require a mechanism to ensure we don't keep too many strings etc.