@staabm 2016-07-27
@nikic thx for the explanation. maybe we could add this capability to interned strings only (and add a flag to those or something). interned strings which tend to get re-used very often would benefit most from such a optimization anywhere, right?