@nikic 2016-07-27
@staabm Interned strings likely wouldn't help for this particular situation though. Furthermore it would require doing an UTF-8 check on all interned strings upfront, as they live in protected memory.

For non-interned strings, adding a (GC) flag directly on the string would be a possibility, but requires care to ensure that this flag is always reset if the string is modified. If we integrate this with forget_hash_val() it might work with little code modification. It feels ugly to include PCRE concerns into the string flags though :/