@staabm 2016-07-27
> It feels ugly to include PCRE concerns into the string flags though :/

agree. I guess it would only make sense in case it makes a big difference in pcre perf.
no sure whether a "this interned string is valid utf8" flag could make sense for other areas.