@nikic 2016-07-27
Also note that we'd still have to make sure that the position is valid -- but as UTF-8 is a self-synchronizing encoding this is cheap.