@ju1ius 2016-07-27
Well, sorry guys, from this point I can't help anymore. The only little knowledge I have  of the internals come from @nikic's (great) blog posts... There should be more of them! :wink: 

As you might have guessed, the reason I submitted this PR is for matching a **lot** of times (I mean tens of thousands of calls) against moderately large inputs (up to a few hundreds of kilobytes) for lexical analysis.
Non-utf8 `preg_match` calls (without `/u`) allow to do that in a reasonable amount of time, but with the quadratic complexity imposed by repeated utf8 validation, it's just ridiculously slow.

I tried to use the `mb_ereg_search` api, which does just that: validating once in `mb_ereg_search_init` (using `php_mb_check_encoding` [under the hood](https://github.com/php/php-src/blob/master/ext/mbstring/php_mbregex.c#L1382)) so that you can call `mb_ereg_search` repeatedly at very a little cost. It's even slightly faster than non-utf8 `preg_match` in my benchmarks.
Unfortunately the Oniguruma integration is kinda flaky. I found a number of little bugs that render it unusable for what I need ATM. :broken_heart: 

So I sincerely hope somebody can come up with some wizardry (or redesign the whole preg utf8 api :laughing:) to make this happen!
In the meantime I'll file bugs against the mbstring extension... :pray:

Have a nice day! :sunny: